Split daemon service code to _windows file
This moves some of the code that was conditionally
executed on Windows to a separate, windows-only file.
