+++ cmd/dockerd/service_unsupported.go
@@ -6,9 +6,5 @@ import (
 	"github.com/spf13/pflag"
 )
 
-func initService(daemonCli *DaemonCli) (bool, bool, error) {
-	return false, false, nil
-}
-
 func installServiceFlags(flags *pflag.FlagSet) {
 }
